feat(backend): start converting exporter to self contained function
